Remove ineffective domainControllerSwizzleGuardForSwizzledObject:…
The key being used for the associated object comes from a pointer to a new object on every pass through the method. Since the key is always changing, we’ll never be able to retrieve a previously set associated object. The check in the if statement will always pass, effectively making this method a no-op.